# Swagger/documentation.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api animalerie",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "sandbox-server"
        }
    ],
    "paths": {
...
        "/produit": {
            "get": {
                "tags": [
                    "Produit"
                ],
                "summary": "List all produits",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List all produits",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Produit[]",
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Produit"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/{id}": {
            "get": {
                "tags": [
                    "Produit"
                ],
                "summary": "Retourne un produit",
                "description": "Retourne un produit",
                "operationId": "abcbc550c1a09c8c10800255187236f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à récupérer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retourne un produit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Produit"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "apiKey missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès à la page refusé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "access key manquante",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/create": {
            "post": {
                "tags": [
                    "Produit"
                ],
                "summary": "Create produit",
                "description": "Créer un produit",
                "operationId": "ddf41842d908a155aefc396a46453626",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Produit à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "laisse"
                                    },
                                    "prix": {
                                        "type": "float",
                                        "example": 19.99
                                    },
                                    "type_animal": {
                                        "type": "string",
                                        "example": "chien"
                                    },
                                    "id_categorie": {
                                        "type": "int",
                                        "example": 3
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/update/{id}": {
            "put": {
                "tags": [
                    "Produit"
                ],
                "summary": "Update produit",
                "description": "Update produit in Db",
                "operationId": "9acdfb2ad70c64bedf5e401eee5ed373",
                "parameters": [
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Produit à enregistrer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "laisse"
                                    },
                                    "prix": {
                                        "type": "float",
                                        "example": 19.99
                                    },
                                    "type_animal": {
                                        "type": "string",
                                        "example": "chien"
                                    },
                                    "id_categorie": {
                                        "type": "int",
                                        "example": 3
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": null
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit enregistré",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/produit/delete/{id}": {
            "delete": {
                "tags": [
                    "Produit"
                ],
                "summary": "Delete produit",
                "description": "Delete produit in Db",
                "operationId": "d1cc4c0e1aff403b3314f9a7f3133fac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id du produit à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "access key",
                        "in": "query",
                        "description": "access key permettant de valider l'application",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Suppression validée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "Produit supprimé"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Produit": {
                "title": "Produit",
                "properties": {
                    "id": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "nom": {
                        "nullable": false,
                        "type": "string"
                    },
                    "prix": {
                        "nullable": false,
                        "type": "float"
                    },
                    "type_animal": {
                        "nullable": false,
                        "type": "string"
                    },
                    "quantite": {
                        "nullable": false,
                        "type": "integer"
                    },
                    "description": {
                        "nullable": true,
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    }
}